Fixed issue in Android API >= 25 causing widget listview to scroll to top when unintended.
This seems to be caused by behaviour changes in appwidget which are causing the setScrollPosition value
to be persisted across remote view updates. Since this value was only ever set to 0 when a widget feed
refresh was performed, this caused the unintended scrolling behaviour after the first refresh of the feed.
To work around this I replaced setScrollPosition(0) to setRelativeScrollPosition(-1000000) and added
calls to reset the relative position before each call to  notifyAppWidgetViewDataChanged.